+++ python/lsst/sims/catUtils/exampleCatalogDefinitions/phoSimCatalogExamples.py
@@ -1,20 +1,80 @@
 """Instance Catalog"""
 import numpy
 from lsst.sims.utils import SpecMap, defaultSpecMap
-from lsst.sims.catalogs.measures.instance import InstanceCatalog
-from lsst.sims.utils import arcsecFromRadians
+from lsst.sims.catalogs.measures.instance import InstanceCatalog, compound
+from lsst.sims.utils import arcsecFromRadians, _observedFromICRS
 from lsst.sims.catUtils.mixins import AstrometryStars, AstrometryGalaxies, \
-                                      EBVmixin
+                                      AstrometrySSM, EBVmixin
 
-__all__ = ["PhosimInputBase", "PhoSimCatalogPoint", "PhoSimCatalogZPoint",
-           "PhoSimCatalogSersic2D", "PhoSimSpecMap"]
+from lsst.sims.utils import sphericalFromCartesian, cartesianFromSpherical
+from lsst.sims.utils import rotationMatrixFromVectors
+
+__all__ = ["PhosimInputBase","PhoSimAstrometryBase",
+           "PhoSimCatalogPoint", "PhoSimCatalogZPoint",
+           "PhoSimCatalogSersic2D", "PhoSimCatalogSSM", "PhoSimSpecMap"]
 
 
 PhoSimSpecMap = SpecMap(fileDict=defaultSpecMap.fileDict,
                         dirDict={'(^lte)':'starSED/phoSimMLT'})
 
 
-class PhosimInputBase(InstanceCatalog):
+class PhoSimAstrometryBase(object):
+
+    def _dePrecess(self, ra_in, dec_in, obs_metadata):
+        """
+        Calculate the displacement between the boresite and the boresite
+        corrected for precession, nutation, and aberration (not refraction).
+
+        Convert boresite and corrected boresite to Cartesian coordinates.
+
+        Calculate the rotation matrix to go between those Cartesian vectors.
+
+        Convert [ra_in, dec_in] into Cartesian coordinates.
+
+        Apply the rotation vector to those Cartesian coordinates.
+
+        Convert back to ra, dec-like coordinates
+
+        @param [in] ra_in is a numpy array of RA in radians
+
+        @param [in] dec_in is a numpy array of Dec in radians
+
+        @param [in] obs_metadata is an ObservationMetaData
+
+        @param [out] ra_out is a numpy array of de-precessed RA in radians
+
+        @param [out] dec_out is a numpy array of de-precessed Dec in radians
+        """
+
+        if len(ra_in)==0:
+            return numpy.array([[],[]])
+
+        xyz_bore = cartesianFromSpherical(numpy.array([obs_metadata._pointingRA]),
+                                          numpy.array([obs_metadata._pointingDec]))
+
+        precessedRA, precessedDec = _observedFromICRS(numpy.array([obs_metadata._pointingRA]),
+                                                      numpy.array([obs_metadata._pointingDec]),
+                                                      obs_metadata=obs_metadata, epoch=2000.0,
+                                                      includeRefraction=False)
+
+        xyz_precessed = cartesianFromSpherical(precessedRA, precessedDec)
+
+        norm = numpy.sqrt(numpy.power(xyz_bore[0],2).sum())
+        xyz_bore = xyz_bore/norm
+
+        norm = numpy.sqrt(numpy.power(xyz_precessed[0],2).sum())
+        xyz_precessed = xyz_precessed/norm
+
+        rotMat = rotationMatrixFromVectors(xyz_precessed[0], xyz_bore[0])
+
+        xyz_list = cartesianFromSpherical(ra_in, dec_in)
+
+        xyz_de_precessed = numpy.array([numpy.dot(rotMat, xx) for xx in xyz_list])
+        ra_deprecessed, dec_deprecessed = sphericalFromCartesian(xyz_de_precessed)
+        return numpy.array([ra_deprecessed, dec_deprecessed])
+
+
+class PhosimInputBase(InstanceCatalog, PhoSimAstrometryBase):
     filtMap = dict([(c, i) for i,c in enumerate('ugrizy')])
 
     specFileMap = PhoSimSpecMap
@@ -109,6 +169,13 @@ class PhoSimCatalogPoint(PhosimInputBase, AstrometryStars, EBVmixin):
     transformations = {'raPhoSim':numpy.degrees, 'decPhoSim':numpy.degrees}
 
 
+    @compound('raPhoSim','decPhoSim')
+    def get_phoSimCoordinates(self):
+        """Getter for unrefracted observed RA, Dec as expected by PhoSim"""
+        raObs, decObs = self.observedStellarCoordinates(includeRefraction = False)
+        return self._dePrecess(raObs, decObs, self.obs_metadata)
+
+
 class PhoSimCatalogZPoint(PhosimInputBase, AstrometryGalaxies, EBVmixin):
     catalog_type = 'phoSim_catalog_ZPOINT'
     column_outputs = ['prefix', 'uniqueId','raPhoSim','decPhoSim','phoSimMagNorm','sedFilepath',
@@ -124,6 +191,17 @@ class PhoSimCatalogZPoint(PhosimInputBase, AstrometryGalaxies, EBVmixin):
     delimiter = " "
     spatialModel = "point"
     transformations = {'raPhoSim':numpy.degrees, 'decPhoSim':numpy.degrees}
+
+
+    @compound('raPhoSim','decPhoSim')
+    def get_phoSimCoordinates(self):
+        """Getter for unrefracted observed RA, Dec as expected by PhoSim"""
+        ra = self.column_by_name('raJ2000')
+        dec = self.column_by_name('decJ2000')
+        raObs, decObs = _observedFromICRS(ra, dec, includeRefraction = False, obs_metadata=self.obs_metadata,
+                                          epoch=self.db_obj.epoch)
+
+        return self._dePrecess(raObs, decObs, self.obs_metadata)
 
 
 class PhoSimCatalogSersic2D(PhoSimCatalogZPoint):
@@ -143,3 +221,34 @@ class PhoSimCatalogSersic2D(PhoSimCatalogZPoint):
     spatialModel = "sersic2d"
     transformations = {'raPhoSim':numpy.degrees, 'decPhoSim':numpy.degrees, 'positionAngle':numpy.degrees,
     'majorAxis':arcsecFromRadians, 'minorAxis':arcsecFromRadians}
+
+
+    @compound('raPhoSim','decPhoSim')
+    def get_phoSimCoordinates(self):
+        """Getter for unrefracted observed RA, Dec as expected by PhoSim"""
+        ra = self.column_by_name('raJ2000')
+        dec = self.column_by_name('decJ2000')
+        raObs, decObs = _observedFromICRS(ra, dec, includeRefraction = False, obs_metadata=self.obs_metadata,
+                                          epoch=self.db_obj.epoch)
+
+        return self._dePrecess(raObs, decObs, self.obs_metadata)
+
+
+class PhoSimCatalogSSM(PhosimInputBase, AstrometrySSM):
+    catalog_type = 'phoSim_catalog_SSM'
+    column_outputs = ['prefix', 'uniqueId','raPhoSim','decPhoSim','phoSimMagNorm','sedFilepath',
+                      'redshift','shear1','shear2','kappa','raOffset','decOffset',
+                      'spatialmodel','galacticExtinctionModel', 'internalExtinctionModel']
+    default_columns = [('redshift', 0., float),('shear1', 0., float), ('shear2', 0., float),
+                       ('kappa', 0., float), ('raOffset', 0., float), ('decOffset', 0., float),
+                       ('galacticExtinctionModel', 'none', (str,3)),
+                       ('internalExtinctionModel', 'none', (str,4))]
+    default_formats = {'S':'%s', 'f':'%.9g', 'i':'%i'}
+    delimiter = " "
+    spatialModel = "point"
+    transformations = {'raPhoSim':numpy.degrees, 'decPhoSim':numpy.degrees}
+
+    @compound('raPhoSim', 'decPhoSim')
+    def get_phoSimCoordinates(self):
+        raObs, decObs = self.observedSSMCoordinates(includeRefraction = False)
+        return self._dePrecess(raObs, decObs, self.obs_metadata)
