+++ python/lsst/sims/catUtils/mixins/AstrometryMixin.py
@@ -15,69 +15,11 @@ from lsst.sims.utils import _observedFromICRS, _pupilCoordsFromRaDec
 from lsst.sims.coordUtils.CameraUtils import chipNameFromPupilCoords, pixelCoordsFromPupilCoords
 from lsst.sims.coordUtils.CameraUtils import focalPlaneCoordsFromPupilCoords
 
-from lsst.sims.utils import sphericalFromCartesian, cartesianFromSpherical
-from lsst.sims.utils import rotationMatrixFromVectors
-
 __all__ = ["AstrometryBase", "AstrometryStars", "AstrometryGalaxies",
            "AstrometrySSM", "CameraCoords"]
 
 class AstrometryBase(object):
     """Collection of astrometry routines that operate on numpy arrays"""
-
-
-    def _dePrecess(self, ra_in, dec_in, obs_metadata):
-        """
-        Calculate the displacement between the boresite and the boresite
-        corrected for precession, nutation, and aberration (not refraction).
-
-        Convert boresite and corrected boresite to Cartesian coordinates.
-
-        Calculate the rotation matrix to go between those Cartesian vectors.
-
-        Convert [ra_in, dec_in] into Cartesian coordinates.
-
-        Apply the rotation vector to those Cartesian coordinates.
-
-        Convert back to ra, dec-like coordinates
-
-        @param [in] ra_in is a numpy array of RA in radians
-
-        @param [in] dec_in is a numpy array of Dec in radians
-
-        @param [in] obs_metadata is an ObservationMetaData
-
-        @param [out] ra_out is a numpy array of de-precessed RA in radians
-
-        @param [out] dec_out is a numpy array of de-precessed Dec in radians
-        """
-
-        if len(ra_in)==0:
-            return numpy.array([[],[]])
-
-        xyz_bore = cartesianFromSpherical(numpy.array([obs_metadata._pointingRA]),
-                                          numpy.array([obs_metadata._pointingDec]))
-
-        precessedRA, precessedDec = _observedFromICRS(numpy.array([obs_metadata._pointingRA]),
-                                                      numpy.array([obs_metadata._pointingDec]),
-                                                      obs_metadata=obs_metadata, epoch=2000.0,
-                                                      includeRefraction=False)
-
-        xyz_precessed = cartesianFromSpherical(precessedRA, precessedDec)
-
-        norm = numpy.sqrt(numpy.power(xyz_bore[0],2).sum())
-        xyz_bore = xyz_bore/norm
-
-        norm = numpy.sqrt(numpy.power(xyz_precessed[0],2).sum())
-        xyz_precessed = xyz_precessed/norm
-
-        rotMat = rotationMatrixFromVectors(xyz_precessed[0], xyz_bore[0])
-
-        xyz_list = cartesianFromSpherical(ra_in, dec_in)
-
-        xyz_de_precessed = numpy.array([numpy.dot(rotMat, xx) for xx in xyz_list])
-        ra_deprecessed, dec_deprecessed = sphericalFromCartesian(xyz_de_precessed)
-        return numpy.array([ra_deprecessed, dec_deprecessed])
-
 
     @compound('glon','glat')
     def get_galactic_coords(self):
@@ -157,8 +99,6 @@ class AstrometryGalaxies(AstrometryBase):
 
     raObserved, decObserved -- the result of applying precession, nutation, aberration, and refraction
     to raICRS, decICRS
-
-    raPhoSim, decPhoSim -- the same as raObserved, decObserved but neglecting refraction
     """
 
     @compound('raICRS', 'decICRS')
@@ -166,17 +106,6 @@ class AstrometryGalaxies(AstrometryBase):
         """Getter for RA, Dec in the International Celestial Reference System with effects
         due to proper motion and radial velocity applied"""
         return numpy.array([self.column_by_name('raJ2000'), self.column_by_name('decJ2000')])
-
-
-    @compound('raPhoSim','decPhoSim')
-    def get_phoSimCoordinates(self):
-        """Getter for unrefracted observed RA, Dec as expected by PhoSim"""
-        ra = self.column_by_name('raJ2000')
-        dec = self.column_by_name('decJ2000')
-        raObs, decObs = _observedFromICRS(ra, dec, includeRefraction = False, obs_metadata=self.obs_metadata,
-                                          epoch=self.db_obj.epoch)
-
-        return self._dePrecess(raObs, decObs, self.obs_metadata)
 
 
     @compound('raObserved','decObserved')
@@ -199,8 +128,6 @@ class AstrometryStars(AstrometryBase):
 
     raObserved, decObserved -- the result of applying precession, nutation, aberration, parallax,
     and refraction to raICRS, decICRS
-
-    raPhoSim, decPhoSim -- the same as raObserved, decObserved but neglecting refraction
     """
 
     def observedStellarCoordinates(self, includeRefraction = True):
@@ -225,12 +152,6 @@ class AstrometryStars(AstrometryBase):
                      includeRefraction = includeRefraction, obs_metadata=self.obs_metadata,
                      epoch=self.db_obj.epoch)
 
-
-    @compound('raPhoSim','decPhoSim')
-    def get_phoSimCoordinates(self):
-        """Getter for unrefracted observed RA, Dec as expected by PhoSim"""
-        raObs, decObs = self.observedStellarCoordinates(includeRefraction = False)
-        return self._dePrecess(raObs, decObs, self.obs_metadata)
 
     @compound('raObserved','decObserved')
     def get_observedCoordinates(self):
@@ -274,10 +195,6 @@ class AstrometrySSM(AstrometryBase):
         return _observedFromICRS(ra, dec, includeRefraction=includeRefraction,
                                  obs_metadata=self.obs_metadata, epoch=self.db_obj.epoch)
 
-    @compound('raPhoSim', 'decPhoSim')
-    def get_phoSimCoordinates(self):
-        raObs, decObs = self.observedSSMCoordinates(includeRefraction = False)
-        return self._dePrecess(raObs, decObs, self.obs_metadata)
 
     @compound('raObserved', 'decObserved')
     def get_observedCoordinates(self):
